refactor(ui): move the keymap tab footer out of the toolbar module
