+++ src/renderer/components/editors/KeymapEditor.tsx
@@ -9,7 +9,8 @@ import { ViewMatrixPanel } from './ViewMatrixPanel'
 import type { KeymapEditorProps as Props } from './keymap-editor-types'
 import { PANEL_COLLAPSED_WIDTH } from './keymap-editor-types'
 export type { KeymapEditorHandle } from './keymap-editor-types'
-import { KeymapToolbar, ViewMatrixZoomRow, useKeymapTabFooter } from './keymap-editor-toolbar'
+import { KeymapToolbar, ViewMatrixZoomRow } from './keymap-editor-toolbar'
+import { useKeymapTabFooter } from './use-keymap-tab-footer'
 import { PopoverForState, popoverInstanceKey } from './keymap-editor-popover'
 import { useInputModes } from './useInputModes'
 import { useKeymapMultiSelect } from './useKeymapMultiSelect'
